FinSync-main/python_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langgraph==0.0.62
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.3
openpyxl==3.1.2
PyMuPDF==1.23.8
Pillow==10.1.0
aiofiles==23.2.1